Separating the assertGC into own file
