+++ lib/java/test-utils/src/main/java/org/enso/test/utils/MemoryUtils.java
@@ -1,0 +1,64 @@
+package org.enso.test.utils;
+
+import static org.junit.Assert.fail;
+
+import java.lang.ref.Reference;
+import java.util.ArrayList;
+
+/** Package private now. Use ContextUtils to turn this check on. */
+final class MemoryUtils {
+  private MemoryUtils() {}
+
+  static void assertGC(String msg, boolean expectGC, Reference<?> ref) {
+    var memory = expectGC ? new ArrayList<>() : null;
+    try {
+      for (var i = 1L; ; i *= 2) {
+        var size = (int) Math.min(i, Integer.MAX_VALUE / 2);
+        if (checkAndAlloc(ref, memory, size)) {
+          break;
+        }
+      }
+    } catch (OutOfMemoryError err) {
+      // launch the JVM with
+      //   -XX:+HeapDumpOnOutOfMemoryError
+      //   -XX:HeapDumpPath=/tmp
+      err.printStackTrace();
+    }
+    assertReference(ref, expectGC, msg, memory);
+  }
+
+  private static void assertReference(
+      Reference<?> ref, boolean expectGC, String msg, ArrayList<Object> memory) {
+    var obj = ref.get();
+    if (expectGC) {
+      if (obj != null) {
+        fail(msg + " ref still alive: " + obj);
+      }
+    } else {
+      if (obj == null) {
+        fail(msg + " ref has been cleaned: " + obj);
+      }
+    }
+    System.getLogger("assertGC").log(System.Logger.Level.TRACE, "Cannot GC {0}", memory);
+  }
+
+  /**
+   * Checks whether {@code ref} has been GCed. If not, it allocates additional memory of given size
+   * to create a "presure" on GC to try harder to GC.
+   *
+   * @param ref
+   * @param memory
+   * @param toAllocate
+   * @return
+   */
+  private static boolean checkAndAlloc(Reference<?> ref, ArrayList<Object> memory, int toAllocate) {
+    if (ref.get() == null) {
+      return true;
+    }
+    System.gc();
+    if (memory != null) {
+      memory.add(new byte[toAllocate]);
+    }
+    return false;
+  }
+}
